==> jupyter-notebooks/step1_Build_RareX_Layer.ipynb <==
# Build the Rare-X layer: disease - patient - symptom associations

This code extracts disease - patient and patient - symptom associations from the Rare-X anonymized dataset (file `data/Survey_Symptoms_US.tsv`). The layer contains three types of nodes: disease, patient and symptom nodes, and two types of edges: disease-patient and patient-symptom edges. The edges are weighted:
* Each connection between diseases and patients carries a uniform weight of 1.
* Relationships between patients and CSHQ symptoms are assigned weights based on the normalized CSHQ score.
* When dealing with binary patient-symptom associations (presence or absence), only symptoms marked as present are considered, each assigned a weight of 1.

The code generates the following layer: ```network/multilayer/RareX_layer.tsv```


## 1) Load data

# import modules
import pandas as pd
import numpy as np
import os

# read the data file
df = pd.read_csv('../data/Survey_Symptoms_US.tsv',sep='\t', header=0, index_col=0)
df.head()

# compute number of diseases, patients, symptoms and scores
nb_diseases = (len(df["Disease Name"].unique()))
print(f"{nb_diseases} diseases")
print(f"{df.shape[0]} patients")
print(f"{len(df.columns[0:214])} symptoms")
print(f"{len(df.columns[214:222])} CSHQ scores")

## 2) Store disease-patient associations in dictionary

Disease-patient associations are stored into a dictionnary with the function `build_dico_diseases_patients`.

def build_dico_diseases_patients(df: pd.DataFrame) -> dict():
    """
    This function creates a dictionary of the diseases described in the 
    Rare-X data file, and their associated patients

    Args:
        df (pd.DataFrame): dataframe containing the loaded Rare-X
        data file
    
    Returns:
        dict: a dictionay with the keys being the diseases described
        in the data file and the values being the patients associated
        with each disease
    """
    dico_diseases_patients = dict()
    i = 0
    for index, row in df.iterrows():
        if not row["Disease Name"] in dico_diseases_patients.keys():
            dico_diseases_patients[str(row["Disease Name"])] = [index]
        else:
            dico_diseases_patients[str(row["Disease Name"])] += [index]
        i += 1
    return(dico_diseases_patients)     

dico_diseases_patients = build_dico_diseases_patients(df=df)

## 3) Create a mapping functions to map the Rare-X diseases names to Orphanet names

We want to map the Rare-X diseases to their corresponding Orphanet names. This mapping has been done manually and stored in file `data/Diseases_Rx_orpha_corres.csv`. We parse this file in order to create a dictionary, using the function `create_mapping_file_diseases_names`, defined below. 

Note that only 15 of the 27 Rare-X diseases can be mapped to Orphanet disease names.

def create_mapping_file_diseases_names(mapping_file: str) -> dict:
    """This function creates a dictionary used for the mapping
    of diseases names in the Rare-X data file to Orphanet names

    Args:
        mapping_file (str): the file containing the mapping of
        diseases idenfifiers

    Returns:
        dict: a dictionary with keys being the diseases described in 
        the Rare-X data file and the values being their mapping 
        according to ORPHANET. If there is no corresponging 
        Orphanet identifier, there is no mapping and the value
        in the dictionary is set to "None"
    """
    dico_mapping = dict()
    mapping = pd.read_csv(mapping_file, sep=";", header=0)
    for index, row in mapping.iterrows():
        dico_mapping[str(row[0])] = row[1]
    return dico_mapping

dico_mapping = create_mapping_file_diseases_names(mapping_file="../data/Diseases_Rx_orpha_corres.csv")
for disease_rarex, disease_orpha in dico_mapping.items():
    print(f"{disease_rarex} : {disease_orpha} \n")

We also define a `map_disease_name` function to change Rare-X disease names into Orphanet disease names in the Rare-X layer, using the dictionary previously created. This function will be called latter, when actually building the Rare-X layer (see function `build_disease_patient_network` defined in section `4` of this notebook).

def map_disease_name(name_to_map: str, mapping_dict: dict) -> str:
    """This function maps a disease name from the Rare-X data file
    to its corresponding Orphanet name from the mapping dictionary

    Args:
        name_to_map (str): the name of the disease to map
        mapping_dict (dict): the mapping dictionary

    Returns:
        str: the ORPAHNET name of the disease if it exists, the 
        original name else
    """
    return mapping_dict[name_to_map] if mapping_dict[name_to_map] != "None" else name_to_map

## 4) Build the Patient-Disease network

Now that we have:
* patient - *Rare-X diseases* associations in variable `dico_diseases_patients`,
* *Rare-X diseases* - *Orphanet diseases* correspondance in variable `dico_mapping`,
we can construct the patient-disease networkd using the function `build_disease_patient_network`. The output is a pandas dataframe containing all patient-disease associations.

Note that:
* For *Rare-X diseases* having a mapping in Orphanet we use the *Orphanet disease* name, instead of the *Rare-X disease* name in the Rare-X layer.

def build_disease_patient_network(dico_diseases_patients: dict, mapping_dict: dict):
    """This function allows to build the network of disease-patient associations.

    Args:
        dico_diseases_patients (dict): the dictionary of disease-patient associations
        mapping_dict (dict): the mapping dictionary of Rare-X disease names to Orphanet disease names
    
    Returns:
        Pandas dataframe
    """
    network = pd.DataFrame(columns=["Source", "Target"])
    i = 0
    for disease in dico_diseases_patients.keys():
        disease_orpha_name = map_disease_name(name_to_map=disease, mapping_dict=mapping_dict)
        for patient in dico_diseases_patients[disease]:
            network._set_value(i, "Source", disease_orpha_name)
            network._set_value(i, "Target", patient)
            i += 1
    return network

disease_patient = build_disease_patient_network(dico_diseases_patients=dico_diseases_patients, 
                                                mapping_dict=dico_mapping)
disease_patient

We add weights to the disease-patient associations, and set it all to `1`.

disease_patient['Weight'] = 1
disease_patient

## 5) Build the patient-symptom network

Now, we aim to create a patient-symptom network. Within this process, the `NA` and any absent values are treated as indicators of the symptom's absence. Conversely, associations between patients and symptoms are assigned a weight of `1` for symptoms that are detected. Meanwhile, the normalized CSHQ scores are employed as the weighting factors for CSHQ symptoms.

First, we normalize CSHQ scores by max value:

for colname in ['CSHQ Subscale 1: Bedtime Resistance',
                'CSHQ Subscale 2: Sleep onset Delay',
                'CSHQ Subscale 3: Sleep Duration',
                'CSHQ Subscale 4: Sleep Anxiety',
                'CSHQ Subscale 5: Night Wakings',
                'CSHQ Subscale 6: Parasomnias',
                'CSHQ Subscale 7: Sleep Disordered Breathing',
                'CSHQ Subscale 8: Daytime Sleepiness']:
    df[colname] = df[colname]/np.nanmax(df[colname])

Now, we can parse the data frame and extract weighted patient-symptoms associations. For boolean symptoms (present/absent), we only extract symptoms that are present, and set the weight to `1`. For CSHQ symptoms, we use the normalized CSHQ score as a weight. 

# Dropping the disease column
symp_df = df.drop(df.columns[222], axis=1)

# Initialize an empty list to store the results
result_rows = []

for idx, row in symp_df.iterrows():
    for column_name, value in row.items():
        if not np.isnan(value) and value!=0:
            result_rows.append([idx, column_name, value])

# Resulting weighted Patient-Symptom network
patient_symptom = pd.DataFrame(result_rows, columns=['Source', 'Target', 'Weight'])
patient_symptom

## 6) Concatenate the patient-disease associations and the patient-symptom associations

The final Rare-X layer is created by concatenating patient-disease and patient-symptoms associations:

rarex_network = pd.concat([disease_patient, patient_symptom], ignore_index=True)
rarex_network

We save this layer in file ```network/multilayer/RareX_layer.tsv```

os.makedirs("../network/multilayer", exist_ok=True)
rarex_network.to_csv("../network/multilayer/RareX_layer.tsv", sep="\t", header=None, index=False)
