==> jupyter-notebooks/step3_Build_Bipartite_Network.ipynb <==
# Bipartite edges: Rare-X disease - Orphanet disease associations

This code builds the bipartite associations that allows connecting the **Rare-X layer** and the **Orphanet layer**. It contains 15 edges, connecting the 15 Rare-X diseases mapped to Orphanet diseases. The network is stored in file `network/bipartite/bipartite_RareX_Orpha_diseases.tsv`.

import pandas as pd
import os

def build_mapping_dict(diseases_file: str, mapping_file: str) -> tuple[list, dict]:
    """Function that generates a mapping dictionary of the correspondances 
    between Rare-X diseases names and Orphanet diseases names.

    Args:
        diseases_file (str): the path to the file that stores the correspondances
        between diseases names
        mapping_file (str): the path to the file that recapitulates the mapping
        betwen Orphanet diseases codes and Orphanet diseases names

    Returns:
        tuple[list, dict]: a list of diseases in the Rare-X layer, a mapping 
        dictionary with keys being the Rare-X disease names and the values 
        being the Orphanet disease names. 
    """
    diseases = pd.read_csv(diseases_file, sep=";", header=0)
    mapping_data = pd.read_csv(mapping_file, sep="\t", header=None)
    dico_diseases_code = dict()
    mapping_dict = dict()
    list_diseases_in_RARE_X_layer = list()
    # create a dictionary of Orphanet diseases and their codes
    for index, row in mapping_data.iterrows():
        if str(row[1]) not in dico_diseases_code.keys():
            dico_diseases_code[str(row[1])] = row[0]
    # create a list of diseases present in the Rare-X network
    for index, row in diseases.iterrows():
        if row[1] == "None":
            list_diseases_in_RARE_X_layer.append(row[0])
        else:
            list_diseases_in_RARE_X_layer.append(row[1])
    # create a mapping dict with the Orphanet diseases present in
    # the Rare-X network and their associated codes
    for disease in list_diseases_in_RARE_X_layer: 
        if disease in dico_diseases_code.keys():
            disease_name_mapped = dico_diseases_code[disease]
            mapping_dict[disease] = disease_name_mapped
    return list_diseases_in_RARE_X_layer, mapping_dict

(list_diseases_in_RARE_X_layer, mapping_dict) = build_mapping_dict(diseases_file="../data/Diseases_Rx_orpha_corres.csv", mapping_file="../data/OrphaDisease_HPO_extract.tsv")

def build_bipartite(list_diseases: list, mapping_dict: dict, bipartite_name: str) -> None:
    """Function that generates the bipartite network that recapitulates the mapping
    between Rare-X diseases names and Orphanet diseases names

    Args:
        list_diseases (list): a list of diseases in the Rare-X layer = diseases to map
        mapping_dict (dict): the mapping dicionary that recapitulates the mapping
        of Rare-X disease names to Orphanet disease names
        bipartite_name (str): path the network file to generate
    
    Returns: None
    """
    bipartite = pd.DataFrame(columns=["source", "target"])
    i = 0
    for disease in list_diseases:
        if disease in mapping_dict.keys():
            disease_orpha_code = mapping_dict[disease]
            bipartite._set_value(i, "source", disease)
            bipartite._set_value(i, "target", disease_orpha_code)
            i += 1
    print(bipartite)
    bipartite.to_csv(bipartite_name, sep="\t", index=False, header=None)

os.makedirs("../network/bipartite", exist_ok=True)
build_bipartite(list_diseases=list_diseases_in_RARE_X_layer, mapping_dict=mapping_dict,
                bipartite_name="../network/bipartite/bipartite_RareX_Orpha_diseases.tsv")
